Load intel compiler module to avoid linking errors
Previous PR introduced the compilation of NEURON and CoreNEURON with Intel compiler. However without it being loaded in the tests `special` binary fails to build due to linking errors (https://bbpgitlab.epfl.ch/hpc/cvf/-/jobs/342413#L4490).
This isn't an issue in the tests since the `lib[core]nrnmech.so` shared libraries are build without a problem but maybe better to solve this issue in case we see anything unexpected in the future.