Fix CI after CoreNEURON merge into NEURON
After https://github.com/neuronsimulator/nrn/pull/2055 and friends, CoreNEURON is no longer used in standalone mode -- it is always built as part of NEURON.

Consequently we cannot use the `coreneuron` Spack package anymore and must install NEURON once for each CoreNEURON that we want.

Because neuron@develop crept into the deployment, we have to use very verbose specs to avoid it.